% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gogr.R
\name{gog}
\alias{gog}
\title{Send data to a gog server}
\usage{
gog(x, url = "http://localhost:4808/data")
}
\arguments{
\item{x}{a data frame}

\item{url}{the gog /data endpoint to send to}
}
\description{
This function takes takes a data frame and sends it to a gog server.
The gog server is responsible for passing the data to a gog frontend
for visualization.
}
\examples{
\dontrun{
gog(iris)
}
}

