% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gogr.R
\docType{package}
\name{gogr}
\alias{gogr}
\alias{gogr-package}
\title{Use gog visualization system from R}
\description{
Communicate with a \href{https://github.com/ajschumacher/gog}{gog}
server for data visualization independent of R.
}

